add permissions checks

+++ src/main/java/keifer/service/DeckServiceImpl.java
@@ -14,10 +14,12 @@ import keifer.persistence.model.UserEntity;
 import keifer.service.model.CardCondition;
 import keifer.service.model.DeckFormat;
 import lombok.NonNull;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import javax.security.sasl.AuthenticationException;
 import javax.servlet.ServletException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -35,7 +37,7 @@ public class DeckServiceImpl implements DeckService {
     private final CardRepository cardRepository;
     private final CardConverter cardConverter;
     private final TcgService tcgService;
-    private final SessionContextService sessionContextService;
+    private final TokenParsingServiceImpl tokenParsingServiceImpl;
 
     public DeckServiceImpl(@NonNull UserRepository userRepository,
                            @NonNull DeckRepository deckRepository,
@@ -43,14 +45,14 @@ public class DeckServiceImpl implements DeckService {
                            @NonNull CardRepository cardRepository,
                            @NonNull CardConverter cardConverter,
                            @NonNull TcgService tcgService,
-                           @NonNull SessionContextService sessionContextService) {
+                           @NonNull TokenParsingServiceImpl tokenParsingServiceImpl) {
         this.userRepository = userRepository;
         this.deckRepository = deckRepository;
         this.deckConverter = deckConverter;
         this.cardRepository = cardRepository;
         this.cardConverter = cardConverter;
         this.tcgService = tcgService;
-        this.sessionContextService = sessionContextService;
+        this.tokenParsingServiceImpl = tokenParsingServiceImpl;
     }
 
     @Override
@@ -62,7 +64,7 @@ public class DeckServiceImpl implements DeckService {
     @Override
     public List<Deck> getDecks(Long userId) {
 
-//        String test = sessionContextService.getSession().getServletContext().g
+        checkPermissions(userId);
 
         List<Deck> decks = new ArrayList<>();
         decks.add(getDeckOverview(userId));
@@ -74,6 +76,8 @@ public class DeckServiceImpl implements DeckService {
 
     @Override
     public Deck getDeck(Long userId, Long deckId) {
+
+        checkPermissions(userId);
 
         // Deck Overview
         if (deckId == 0) {
@@ -87,6 +91,8 @@ public class DeckServiceImpl implements DeckService {
 
     @Override
     public Card saveCard(Long userId, Long deckId, Card card) {
+
+        checkPermissions(userId);
 
         DeckEntity deckEntity = fetchDeck(deckId);
 
@@ -123,6 +129,8 @@ public class DeckServiceImpl implements DeckService {
     @Override
     public void saveDeck(Long userId, Deck deck) throws ServletException {
 
+        checkPermissions(userId);
+
         DeckEntity deckEntity = null;
 
         if (deck.getId() == null || deck.getId() < 1) {
@@ -147,6 +155,8 @@ public class DeckServiceImpl implements DeckService {
 
     @Override
     public void refreshDeck(Long userId, Long deckId) {
+
+        checkPermissions(userId);
 
         // Deck Overview
         if (deckId == 0) {
@@ -181,11 +191,15 @@ public class DeckServiceImpl implements DeckService {
     @Override
     public void deleteCard(Long userId, Long cardId) {
 
+        checkPermissions(userId);
+
         cardRepository.deleteById(cardId);
     }
 
     @Override
     public void deleteDeck(Long userId, Long deckId) {
+
+        checkPermissions(userId);
 
         DeckEntity deckEntity = fetchDeck(deckId);
 
@@ -193,6 +207,8 @@ public class DeckServiceImpl implements DeckService {
     }
 
     private Deck getDeckOverview(Long userId) {
+
+        checkPermissions(userId);
 
         Deck deck = Deck.builder().id(0L).name("Deck Overview").cards(new ArrayList<>()).build();
 
@@ -264,9 +280,17 @@ public class DeckServiceImpl implements DeckService {
         deckRepository.save(deckEntity);
     }
 
+    @SneakyThrows
+    private void checkPermissions(Long id) {
+
+        if (!id.equals(tokenParsingServiceImpl.getUserId())) {
+            throw new AuthenticationException("User is not authorized.");
+        }
+    }
+
     // Fires at 12 PM every day
     @Scheduled(cron = "0 0 12 * * ?")
-    private void refreshAllDecks() {
+    public void refreshAllDecks() {
 
         System.out.println("Scheduled task running.");
